pyproject.toml
[project]
name = "digital-dna"
version = "0.1.0"
description = "Build a personality knowledge graph from founder content and generate authentic social posts"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "spacy>=3.7.0",
    "textstat>=0.7.0",
    "pandas>=2.0.0",
    "ollama>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "scikit-learn>=1.3.0",
    "streamlit>=1.30.0",
    "pyvis>=0.3.0",
    "streamlit-agraph>=0.0.45",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
]

[project.scripts]
digital-dna = "src.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]
